requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-cors==0.7.0
aiosignal==1.3.1
alembic==1.14.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
axial_positional_embedding==0.3.12
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorful==0.5.6
colorlog==6.9.0
CoLT5-attention==0.11.1
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.9
einops==0.8.0
filelock==3.13.1
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.2.0
google-api-core==2.24.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
huggingface-hub==0.27.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
local-attention==1.10.0
lxml==5.3.1
Mako==1.3.8
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
obspy==1.4.1
opencensus==0.11.4
opencensus-context==0.1.3
optuna==4.2.0
pandas==2.0.3
path==17.0.0
patsy==1.0.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
product_key_memory==0.2.11
prometheus_client==0.21.1
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.3
py-spy==0.4.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
pytz==2024.2
PyYAML==6.0.2
ray==2.10.0
referencing==0.35.1
reformer-pytorch==1.4.4
requests==2.32.3
rpds-py==0.20.1
rsa==4.9
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
smart-open==7.1.0
SQLAlchemy==1.4.54
statsmodels==0.14.1
sympy==1.13.1
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tqdm==4.67.1
triton==3.0.0
tzdata==2025.1
urllib3==2.2.3
virtualenv==20.29.1
wrapt==1.17.2
yarl==1.15.2
zipp==3.20.2
